Fix Unit Test
